BugFix: carousel component Image size 설정 & navbar Logo priority 설정 -> 경고 제거
